test: re-organized Cypress folder structure
