+++ gravitee-apim-cypress/cypress/integration/api/visibility.spec.ts
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ADMIN_USER, API_PUBLISHER_USER, APPLICATION_USER, LOW_PERMISSION_USER } from '../../fakers/users/users';
-import { ApiFakers } from '../../fakers/apis';
-import { Api, ApiErrorCodes, ApiLifecycleState, ApiMember, ApiVisibility, PortalApi } from '../../model/apis';
-import { CollectionResponse, PortalError } from '../../model/technical';
-import { User } from '../../model/users';
-import { ApiAssertions, PortalApiAssertions } from '../../assertions/api.assertion';
-import { ErrorAssertions } from '../../assertions/error.assertion';
-import { gio } from '../../commands/gravitee.commands';
+import { ADMIN_USER, API_PUBLISHER_USER, APPLICATION_USER, LOW_PERMISSION_USER } from '../../fixtures/users/users.fixtures';
+import { ApiFakers } from '../../support/utils/faker_apis';
+import { Api, ApiErrorCodes, ApiLifecycleState, ApiMember, ApiVisibility, PortalApi } from '../../support/model/apis';
+import { CollectionResponse, PortalError } from '../../support/model/technical';
+import { User } from '../../support/model/users';
+import { ApiAssertions, PortalApiAssertions } from '../../support/assertions/api.assertion';
+import { ErrorAssertions } from '../../support/assertions/error.assertion';
+import { gio } from '../../support/utils/gravitee.commands';
 
 context('API - Visibility', () => {
   describe('Private', () => {
