+++ gravitee-apim-cypress/cypress/integration/bulk/cypress-config.json
@@ -1,6 +1,6 @@
 {
   "baseUrl": "http://localhost:8083",
-  "integrationFolder": "cypress/bulk",
+  "integrationFolder": "cypress/tests/bulk",
   "testFiles": "**/*.ts",
   "video": false
 }
