+++ gravitee-apim-cypress/cypress/integration/ui/ui-api-list.spec.ts
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ADMIN_USER, API_PUBLISHER_USER } from '../fakers/users/users';
-import { ApiFakers } from '../fakers/apis';
-import { Api } from '../model/apis';
-import { ApiAssertions } from '../assertions/api.assertion';
-import { gio } from '../commands/gravitee.commands';
+import { ADMIN_USER, API_PUBLISHER_USER } from '../../fixtures/users/users.fixtures';
+import { ApiFakers } from '../../support/utils/faker_apis';
+import { Api } from '../../support/model/apis';
+import { ApiAssertions } from '../../support/assertions/api.assertion';
+import { gio } from '../../support/utils/gravitee.commands';
 
-describe('API List feature', () => {
+describe.skip('API List feature', () => {
   let createdApi: Api;
 
   before(() => {
