+++ gravitee-apim-cypress/cypress/integration/ui/ui-login.spec.ts
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ADMIN_USER } from '../fakers/users/users';
+import { ADMIN_USER } from '../../fixtures/users/users.fixtures';
 
-describe('Login Feature', () => {
+describe.skip('Login Feature', () => {
   // Here we use a beforeEach and an after each to reset properly all the cookie we want
   // otherwise we are sometimes redirected or face XRCF issues
   beforeEach(() => {
