+++ gravitee-apim-cypress/cypress/support/assertions/environment-configuration.assertion.ts
