+++ gravitee-apim-cypress/cypress/support/model/quality-rules.ts
