+++ gravitee-apim-cypress/package.json
@@ -5,11 +5,11 @@
     "description": "Gravitee.io APIM - Cypress",
     "scripts": {
         "test": "cypress open",
-        "test:bulk": "cypress run --config-file cypress/bulk/cypress-config.json",
+        "test:bulk": "cypress run --config-file cypress/integration/bulk/cypress-config.json",
         "test:ci": "cypress run",
-        "test:e2e": "cypress open --config-file cypress/e2e/cypress-config.json",
+        "test:e2e": "cypress open --config-file cypress/integration/cypress-config.json",
         "test:dev": "cypress run --env printLogsToConsole=always",
-        "test:parallel": "cypress run --record --group apis --spec 'cypress/nrt/apis/**/*'",
+        "test:parallel": "cypress run --record --group apis --spec 'cypress/integration/api/**/*'",
         "lint": "echo 'Lint to be configured'",
         "lint:fix": "npm run prettier:fix && npm run lint:license:fix",
         "lint:license": "license-check-and-add check -f license-check-config.json",
